Added useTrending and usePersonalQueues hooks, along with JSDoc comments
